@@ -1,15 +1,17 @@
 module github.com/onmetal/fedhcp
 
-go 1.17
+go 1.19
 
 require (
 	github.com/coredhcp/coredhcp v0.0.0-20220602152301-a2552c5c1b7a
+	github.com/google/addlicense v1.0.0
 	github.com/insomniacslk/dhcp v0.0.0-20211209223715-7d93572ebe8e
 	github.com/sirupsen/logrus v1.9.0
 	github.com/spf13/pflag v1.0.6-0.20201009195203-85dd5c8bc61c
 )
 
 require (
+	github.com/bmatcuk/doublestar/v4 v4.0.2 // indirect
 	github.com/chappjc/logrus-prefix v0.0.0-20180227015900-3a1d64819adb // indirect
 	github.com/fsnotify/fsnotify v1.4.9 // indirect
 	github.com/google/gopacket v1.1.19 // indirect
@@ -29,6 +31,7 @@ require (
 	github.com/u-root/uio v0.0.0-20220204230159-dac05f7d2cb4 // indirect
 	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519 // indirect
 	golang.org/x/net v0.0.0-20211015210444-4f30a5c0130f // indirect
+	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e // indirect
 	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
 	golang.org/x/term v0.0.0-20210916214954-140adaaadfaf // indirect
 	golang.org/x/text v0.3.7 // indirect
